+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.TeamsTenantNetworkRegion.Tests.ps1
@@ -1,0 +1,183 @@
+[CmdletBinding()]
+param(
+)
+$M365DSCTestFolder = Join-Path -Path $PSScriptRoot `
+                        -ChildPath "..\..\Unit" `
+                        -Resolve
+$CmdletModule = (Join-Path -Path $M365DSCTestFolder `
+            -ChildPath "\Stubs\Microsoft365.psm1" `
+            -Resolve)
+$GenericStubPath = (Join-Path -Path $M365DSCTestFolder `
+    -ChildPath "\Stubs\Generic.psm1" `
+    -Resolve)
+Import-Module -Name (Join-Path -Path $M365DSCTestFolder `
+        -ChildPath "\UnitTestHelper.psm1" `
+        -Resolve)
+
+$Global:DscHelper = New-M365DscUnitTestHelper -StubModule $CmdletModule `
+    -DscResource "TeamsTenantNetworkRegion" -GenericStubModule $GenericStubPath
+Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
+    InModuleScope -ModuleName $Global:DscHelper.ModuleName -ScriptBlock {
+        Invoke-Command -ScriptBlock $Global:DscHelper.InitializeScript -NoNewScope
+        BeforeAll {
+
+            $secpasswd = ConvertTo-SecureString "f@kepassword1" -AsPlainText -Force
+            $Credential = New-Object System.Management.Automation.PSCredential ("tenantadmin@mydomain.com", $secpasswd)
+
+            Mock -CommandName Confirm-M365DSCDependencies -MockWith {
+            }
+
+            Mock -CommandName Get-PSSession -MockWith {
+            }
+
+            Mock -CommandName Remove-PSSession -MockWith {
+            }
+
+            Mock -CommandName Set-CsTenantNetworkRegion -MockWith {
+            }
+
+            Mock -CommandName New-CsTenantNetworkRegion -MockWith {
+            }
+
+            Mock -CommandName Remove-CsTenantNetworkRegion -MockWith {
+            }
+
+            Mock -CommandName New-M365DSCConnection -MockWith {
+                return "Credentials"
+            }<AssignmentMock>
+        }
+        # Test contexts
+        Context -Name "The TeamsTenantNetworkRegion should exist but it DOES NOT" -Fixture {
+            BeforeAll {
+                $testParams = @{
+<TargetResourceFakeValues>                    Ensure = "Present"
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-CsTenantNetworkRegion -MockWith {
+                    return $null
+                }
+            }
+            It "Should return Values from the Get method" {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Absent'
+            }
+            It 'Should return false from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+            It 'Should Create the group from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName New-CsTenantNetworkRegion -Exactly 1
+            }
+        }
+
+        Context -Name "The TeamsTenantNetworkRegion exists but it SHOULD NOT" -Fixture {
+            BeforeAll {
+                $testParams = @{
+<TargetResourceFakeValues>                    Ensure = "Absent"
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-CsTenantNetworkRegion -MockWith {
+                    return @{
+                    BypassID              = "FakeStringValue"
+                    Description           = "FakeStringValue"
+                    CentralSite           = "FakeStringValue"
+                    Identity              = "FakeStringValue"
+
+                    }
+                }
+            }
+
+            It "Should return Values from the Get method" {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+            }
+
+            It 'Should return true from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It 'Should Remove the group from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Remove-CsTenantNetworkRegion -Exactly 1
+            }
+        }
+        Context -Name "The TeamsTenantNetworkRegion Exists and Values are already in the desired state" -Fixture {
+            BeforeAll {
+                $testParams = @{
+<TargetResourceFakeValues>                    Ensure = "Present"
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-CsTenantNetworkRegion -MockWith {
+                    return @{
+                    BypassID              = "FakeStringValue"
+                    Description           = "FakeStringValue"
+                    CentralSite           = "FakeStringValue"
+                    Identity              = "FakeStringValue"
+
+                    }
+                }
+            }
+
+
+            It 'Should return true from the Test method' {
+                Test-TargetResource @testParams | Should -Be $true
+            }
+        }
+
+        Context -Name "The TeamsTenantNetworkRegion exists and values are NOT in the desired state" -Fixture {
+            BeforeAll {
+                $testParams = @{
+<TargetResourceFakeValues>                    Ensure = "Present"
+                    Credential = $Credential;
+                }
+
+                Mock -CommandName Get-CsTenantNetworkRegion -MockWith {
+                    return @{
+                    BypassID              = "FakeStringValueDrift #Drift"
+                    Description           = "FakeStringValueDrift #Drift"
+                    CentralSite           = "FakeStringValueDrift #Drift"
+                    Identity              = "FakeStringValue"
+                    }
+                }
+            }
+
+            It "Should return Values from the Get method" {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+            }
+
+            It 'Should return false from the Test method' {
+                Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It "Should call the Set method" {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Set-CsTenantNetworkRegion -Exactly 1
+            }
+        }
+
+        Context -Name "ReverseDSC Tests" -Fixture {
+            BeforeAll {
+                $Global:CurrentModeIsExport = $true
+                $testParams = @{
+                    Credential = $Credential
+                }
+
+                Mock -CommandName Get-CsTenantNetworkRegion -MockWith {
+                    return @{
+                    BypassID              = "FakeStringValue"
+                    Description           = "FakeStringValue"
+                    CentralSite           = "FakeStringValue"
+                    Identity              = "FakeStringValue"
+
+                    }
+                }
+            }
+            It "Should Reverse Engineer resource from the Export method" {
+                Export-TargetResource @testParams
+            }
+        }
+    }
+}
+
+Invoke-Command -ScriptBlock $Global:DscHelper.CleanupScript -NoNewScope