+++ CHANGELOG.md
@@ -5,11 +5,16 @@
 * PlannerTask
   * Fixed issue where Attachments Uri weren't properly exiting single quotes.
     FIXES [#2822](https://github.com/microsoft/Microsoft365DSC/issues/2822)
+* PPPowerAppsEnvironment
+  * Adds support for Developr SKU and fix for Teams SKU
+    FIXES [#2821](https://github.com/microsoft/Microsoft365DSC/issues/2821)
 * SCProtectionAlert
   * Support for certificate based auth
   * Fix removal of alert
   * Added additional supported values for ThreatType, Severity & AggregationType
     FIXES [#2793](https://github.com/microsoft/Microsoft365DSC/issues/2793)
+* TeamsTenantNetworkRegion
+  * Initial Release.
 
 # 1.23.125.1